Revert "[Merge]: remote-tracking branch 'origin/main' into feat/MainMenu"
This reverts commit ee3ec9308e63a737d55411aeb5a1071873977744, reversing
changes made to 8d71d557dcad6bc9ad691a3993abaf45c580ae4c.

+++ Pontiland/src/main/resources/SQL/DDL.sql
@@ -41,7 +41,7 @@ CREATE TABLE Propiedad (
 
 -- Tabla Partida (Representación de la partida en la base de datos)
 CREATE TABLE Partida (
-    PartidaID BIGINT PRIMARY KEY,
+    PartidaID INT PRIMARY KEY,
     Activa BOOLEAN NOT NULL DEFAULT TRUE,
     NumeroJugadores INT NOT NULL CHECK (NumeroJugadores BETWEEN 2 AND 4)
 );
